Added added pagination logic to a helper hook for leaderboard functions
